try to build submission before grading
We were seeing errors when students submitted assignments that added imports not present in the stencil..
Hypothesis: this is because the file, and hence the inputs, were not built.
This change will try to build the submission in advance.
In the process, it improves error handling, noting when the autograder itself doesn't compile (rare) or when the submission doesn't compile.